customer_transaction_prediction_backend/requirements.txt
fastapi
uvicorn
pymongo
pandas
numpy
joblib
python-dotenv
python-multipart
scikit-learn